Сгруппировала types в один authTypes
