--- client/src/modules/auth-form/types/AuthResponse.ts
@@ -1,7 +1,0 @@
-import { TUser } from "./TUser";
-
-export type AuthResponse = {
-    accessToken: string,
-    refreshToken: string,
-    user: TUser,
-}
