--- client/src/modules/auth-form/types/TFormData.ts
@@ -1,6 +1,0 @@
-export type TFormData = {
-    nickname: string;
-    email: string;
-    password: string;
-    confirmPassword: string;
-}
