--- client/src/modules/auth-form/types/TUser.ts
@@ -1,5 +1,0 @@
-export type TUser = {
-    email: string,
-    isActivated: boolean,
-    id: string,
-}
