+++ client/src/modules/auth-form/types/authTypes.ts
@@ -1,0 +1,26 @@
+export type AuthResponse = {
+    accessToken: string,
+    refreshToken: string,
+    user: TUser,
+}
+
+export type TFormData = {
+    nickname: string;
+    email: string;
+    password: string;
+    confirmPassword: string;
+}
+
+export type TState = {
+    error_message: string;
+    is_loader_active: boolean;
+    user: TUser | null;
+    user_auth: boolean;
+    user_reg: boolean;
+}
+
+export type TUser = {
+    email: string,
+    isActivated: boolean,
+    id: string,
+}
